+++ app/controllers/vehicles_controller.rb
@@ -1,5 +1,7 @@
 class VehiclesController < ApplicationController
   before_action :set_vehicle, only: [:show, :edit, :update, :destroy]
+  after_action :verify_authorized, only: [:new, :create, :edit, :destroy]
+
   def index
     # @vehicles =Vehicle.available_vehicles
     # render  json: {status: 'success', vehicles: @vehicles }
@@ -30,19 +32,51 @@ class VehiclesController < ApplicationController
 
   def new
     @vehicle = Vehicle.new 
+    authorize Vehicle
   end 
 
   def search
+  
     # ick_up = params[:pick_up]
     # drop = params[:drop]
     # results = Geocoder.search(params[:pick_up])
     # results.first.coordinates
     @pick_up = params[:pick_up]
-    @vehicles = Vehicle.where("vehicles.location LIKE ? ", ["%#{@pick_up}%"])
+    @drop = params[:drop]
 
+    pick_up = Geocoder.search(params[:pick_up])
+    pick_up_coordinates= pick_up.first.coordinates
+    drop = Geocoder.search(params[:drop])
+    drop_coordinates = drop.first.coordinates
+    @distance = Geocoder::Calculations.distance_between(pick_up_coordinates, drop_coordinates, :units => :km)
+    # pick_up_coordinates = Geocoder.coordinates(params[:pick_up])
+    # drop_coordinates = Geocoder.coordinates(params[:drop])
+    # @booking.distance = Geocoder::Calculations.distance_between(pick_up_coordinates, drop_coordinates)
+    @vehicles = Vehicle.where("vehicles.location LIKE ? ", ["%#{@pick_up}%"])
+    
     # @vehicles = Vehicle.where('location' LIKE? )
     
   end
+
+  
+  # def search
+  #   pick_up =Geocoder.search(params[:pick_up])
+  #   @p = pick_up.first.coordinates 
+  #   @p.save!
+  #   drop =Geocoder.search(params[:drop])
+  #   @d = drop.first.coordinates
+  #   @d.save!
+  #   location = Geocoder.search(params[:location])
+  #   @l = location.first.coordinates
+  #   @l.save!
+  #   if @p.near(@l, 20)
+  #     @vehicles = Vehicle.where("vehicles.location LIKE ? ", ["%#{@pick_up}%"])
+  #   else
+  #     @vehicles.all
+  #   end
+
+
+  # end
 
   
   #   byebug
@@ -55,10 +89,11 @@ class VehiclesController < ApplicationController
   #   end
   # end 
   def edit
-    @vehicle = Vehicle.find(params[:id])
+    authorize Vehicle
   end
 
   def create
+    authorize Vehicle
     @vehicle = Vehicle.new(vehicle_params)
 
     respond_to do |format|
@@ -73,6 +108,7 @@ class VehiclesController < ApplicationController
   end
 
   def update
+    authorize Vehicle
     respond_to do |format|
       if @vehicle.update(vehicle_params)
         format.html { redirect_to @vehicle, notice: "Vehicle was successfully updated." }
@@ -86,7 +122,7 @@ class VehiclesController < ApplicationController
 
   # DELETE /vehicles/1 or /vehicles/1.json
   def destroy
-    
+    authorize Vehicle
     if @vehicle.destroy  
       flash[:notice] = 'Vehicle deleted!'   
       redirect_to root_path   
@@ -106,6 +142,6 @@ class VehiclesController < ApplicationController
 
     # Only allow a list of trusted parameters through.
     def vehicle_params
-      params.require(:vehicle).permit(:id, :vehicle_type, :license_number, :vehicle_model, :location, :status, :image, :lat, :Long, :search)
+      params.require(:vehicle).permit(:id, :vehicle_type, :license_number, :vehicle_model, :location, :status, :image, :lat, :Long)
     end
 end
